passed down props from UPicked.js to UPickedCard.js, and mapped props within the dumb component.
